chore(): update folder structure
