+++ src/components/CheckIn/styles.js
