+++ src/components/NewCheckIn/styles.js
