+++ src/components/PartnerHistory/styles.js
