+++ src/components/PartnerList/styles.js
