+++ src/navigation/rootNavigator.js
@@ -1,9 +1,7 @@
 import { StackNavigator } from 'react-navigation'
 
-import Login from '../components/Login/Login'
+import Login from '../components/Login'
 import drawerNavigator from './drawerNavigator'
-
-import PartnerList from '../components/PartnerList/PartnerList'
 
 export default RootNavigator = StackNavigator({
   Login: {
